Corrected mnist tutorial to use .jld statistics file
The code for mnist.jl used the .jld file type, but the tutorial file
located in /docs/tutorial/mnist.rst had the snapshot file type
as HDF5.  This resulted in a statistics.hdf5 file that inluded
accuracy and obj_val in an HDF5.compound type.  Unfortunately,
reading compound HDF5 types is not documented in the HDF5
julia module.  Therefore, this file type presents a barrier to
beginners that might be coming to Julia.